+++ CHANGELOG.md
@@ -18,6 +18,7 @@ __BEGIN-CHANGELOG__
 ### [2.0.4] - 2023-12-21
 #### Added
  - Tests for handle_commands for both `Message` and `Slash` objects
+ - Admin-only command support
 #### Changed
  - When slash commands are used, try to respond in ephemeral, noting that we can't use slash in threads.
 #### Fixed